+++ langserver/handle_text_document_formatting.go
@@ -3,6 +3,12 @@ package langserver
 import (
 	"context"
 	"encoding/json"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
 
 	"github.com/sourcegraph/jsonrpc2"
 )
@@ -17,5 +23,79 @@ func (h *langHandler) handleTextDocumentFormatting(ctx context.Context, conn *js
 		return nil, err
 	}
 
-	return h.formatFile(params.TextDocument.URI)
+	return h.formatting(params.TextDocument.URI)
+}
+
+func (h *langHandler) formatting(uri string) ([]TextEdit, error) {
+	f, ok := h.files[uri]
+	if !ok {
+		return nil, fmt.Errorf("document not found: %v", uri)
+	}
+
+	fname, err := fromURI(uri)
+	if err != nil {
+		return nil, fmt.Errorf("invalid uri: %v: %v", err, uri)
+	}
+	fname = filepath.ToSlash(fname)
+	if runtime.GOOS == "windows" {
+		fname = strings.ToLower(fname)
+	}
+
+	configs, ok := h.configs[f.LanguageId]
+	if !ok {
+		configs, ok = h.configs["_"]
+		if !ok || len(configs) < 1 {
+			return nil, fmt.Errorf("format for languageId not supported: %v", f.LanguageId)
+		}
+	}
+	found := 0
+	for _, config := range configs {
+		if config.FormatCommand != "" {
+			found++
+		}
+	}
+	if found == 0 {
+		return nil, fmt.Errorf("format for languageId not supported: %v", f.LanguageId)
+	}
+
+	for _, config := range configs {
+		if config.FormatCommand == "" {
+			continue
+		}
+
+		var command string
+
+		if strings.Index(config.FormatCommand, "${INPUT}") != -1 {
+			command = strings.Replace(config.FormatCommand, "${INPUT}", fname, -1)
+		} else {
+			command = config.FormatCommand + " " + fname
+		}
+
+		var cmd *exec.Cmd
+		if runtime.GOOS == "windows" {
+			cmd = exec.Command("cmd", "/c", command)
+		} else {
+			cmd = exec.Command("sh", "-c", command)
+		}
+		cmd.Env = append(os.Environ(), config.Env...)
+		cmd.Stdin = strings.NewReader(f.Text)
+		b, err := cmd.CombinedOutput()
+		if err != nil {
+			continue
+		}
+		h.logger.Println("format succeeded")
+		text := strings.Replace(string(b), "\r", "", -1)
+		flines := strings.Split(f.Text, "\n")
+		return []TextEdit{
+			{
+				Range: Range{
+					Start: Position{Line: 0, Character: 0},
+					End:   Position{Line: len(flines), Character: len(flines[len(flines)-1])},
+				},
+				NewText: text,
+			},
+		}, nil
+	}
+
+	return nil, fmt.Errorf("format for languageId not supported: %v", f.LanguageId)
 }
